+++ src/tests/api/test_api_views.py
@@ -292,6 +292,7 @@ def test_orga_can_see_all_speakers(
         'submissions',
         'answers',
         'avatar',
+        'availabilities',
     }
     assert set(content['results'][0]['answers'][0].keys()) == {
         'answer',